+++ src/main/java/com/caffeine/gwanghwamun/domain/ai/controller/AiController.java
@@ -3,8 +3,8 @@ package com.caffeine.gwanghwamun.domain.ai.controller;
 import com.caffeine.gwanghwamun.common.response.ApiResponse;
 import com.caffeine.gwanghwamun.common.response.ResponseUtil;
 import com.caffeine.gwanghwamun.common.success.SuccessCode;
-import com.caffeine.gwanghwamun.domain.ai.dto.AiResDTO;
-import com.caffeine.gwanghwamun.domain.ai.dto.AiUpdateReqDTO;
+import com.caffeine.gwanghwamun.domain.ai.dto.response.AiResDTO;
+import com.caffeine.gwanghwamun.domain.ai.dto.request.AiUpdateReqDTO;
 import com.caffeine.gwanghwamun.domain.ai.dto.request.AiPromptReqDTO;
 import com.caffeine.gwanghwamun.domain.ai.service.AiService;
 import io.swagger.v3.oas.annotations.Operation;
