+++ src/main/java/com/caffeine/gwanghwamun/domain/ai/service/AiService.java
@@ -2,8 +2,8 @@ package com.caffeine.gwanghwamun.domain.ai.service;
 
 import com.caffeine.gwanghwamun.common.exception.CustomException;
 import com.caffeine.gwanghwamun.common.exception.ErrorCode;
-import com.caffeine.gwanghwamun.domain.ai.dto.AiResDTO;
-import com.caffeine.gwanghwamun.domain.ai.dto.AiUpdateReqDTO;
+import com.caffeine.gwanghwamun.domain.ai.dto.response.AiResDTO;
+import com.caffeine.gwanghwamun.domain.ai.dto.request.AiUpdateReqDTO;
 import com.caffeine.gwanghwamun.domain.ai.entity.Ai;
 import com.caffeine.gwanghwamun.domain.ai.repository.AiRepository;
 import com.google.genai.Client;
